requirements.txt
fastapi~=0.109.2
requests~=2.31.0
uvicorn==0.27.1
python-multipart~=0.0.9
certifi==2024.2.2
PyOpenSSL==24.1.0